+++ Yeti/Controllers/FeedVC/FeedVC+ContextMenus.m
@@ -28,7 +28,9 @@
         
         if (item.isRead == YES) {
             
-            read = [UIAction actionWithTitle:@"Unread" image:[UIImage systemImageNamed:@"circle"] identifier:nil handler:^(__kindof UIAction * _Nonnull action) {
+            read = [UIAction actionWithTitle:@"Mark Unread" image:[UIImage systemImageNamed:@"circle"] identifier:nil handler:^(__kindof UIAction * _Nonnull action) {
+                
+                [MyFeedsManager article:item markAsRead:NO];
                 
                 [self userMarkedArticle:item read:NO];
                 
@@ -36,7 +38,9 @@
             
         }
         else {
-            read = [UIAction actionWithTitle:@"Read" image:[UIImage systemImageNamed:@"largecircle.fill.circle"] identifier:nil handler:^(__kindof UIAction * _Nonnull action) {
+            read = [UIAction actionWithTitle:@"Mark Read" image:[UIImage systemImageNamed:@"largecircle.fill.circle"] identifier:nil handler:^(__kindof UIAction * _Nonnull action) {
+                
+                [MyFeedsManager article:item markAsRead:YES];
                 
                 [self userMarkedArticle:item read:YES];
                 
@@ -80,6 +84,22 @@
             
         }];
         
+        NSString *directionalNewerImageName = self.sortingOption == YTSortAllDesc || self.sortingOption == YTSortUnreadDesc ? @"arrow.up.circle.fill" : @"arrow.down.circle.fill";
+        
+        NSString *directionalOlderImageName = self.sortingOption == YTSortAllAsc || self.sortingOption == YTSortUnreadAsc ? @"arrow.up.circle.fill" : @"arrow.down.circle.fill";
+        
+        UIAction *directionalNewer = [UIAction actionWithTitle:@"Mark Newer Read" image:[UIImage systemImageNamed:directionalNewerImageName] identifier:nil handler:^(__kindof UIAction * _Nonnull action) {
+           
+            [self markAllNewerRead:indexPath];
+            
+        }];
+        
+        UIAction *directionalOlder = [UIAction actionWithTitle:@"Mark Older Read" image:[UIImage systemImageNamed:directionalOlderImageName] identifier:nil handler:^(__kindof UIAction * _Nonnull action) {
+           
+            [self markAllOlderRead:indexPath];
+            
+        }];
+        
         if (self.type == FeedVCTypeNatural) {
             
             NSString *author;
@@ -109,13 +129,19 @@
                     
                 }];
                 
-                return [UIMenu menuWithTitle:@"Article Actions" children:@[read, bookmark, browser, share, authorAction]];
+                return [UIMenu menuWithTitle:@"Article Actions" children:@[read, bookmark, browser, share, authorAction, directionalNewer, directionalOlder]];
                 
             }
             
         }
         
-        return [UIMenu menuWithTitle:@"Article Actions" children:@[read, bookmark, browser, share]];
+        if (self.type == FeedVCTypeAuthor || self.type == FeedVCTypeBookmarks || self.type == FeedTypeFolder) {
+            
+            return [UIMenu menuWithTitle:@"Article Actions" children:@[read, bookmark, browser, share, directionalNewer, directionalOlder]];
+            
+        }
+        
+        return [UIMenu menuWithTitle:@"Article Actions" children:@[read, bookmark, browser, share, directionalNewer, directionalOlder]];
         
     }];
     
@@ -139,6 +165,8 @@
         
         read = [UIContextualAction contextualActionWithStyle:UIContextualActionStyleNormal title:@"Unread" handler:^(UIContextualAction * _Nonnull action, __kindof UIView * _Nonnull sourceView, void (^ _Nonnull completionHandler)(BOOL)) {
             
+            [MyFeedsManager article:item markAsRead:NO];
+            
             [self userMarkedArticle:item read:NO];
             
             completionHandler(YES);
@@ -151,6 +179,8 @@
     else {
         
         read = [UIContextualAction contextualActionWithStyle:UIContextualActionStyleNormal title:@"Read" handler:^(UIContextualAction * _Nonnull action, __kindof UIView * _Nonnull sourceView, void (^ _Nonnull completionHandler)(BOOL)) {
+            
+            [MyFeedsManager article:item markAsRead:YES];
             
             [self userMarkedArticle:item read:YES];