+++ changelog.md
@@ -1,5 +1,8 @@
 # v2.0.0
 
+##  Build 16
+ - Directional marking as read. 
+
 ## Build 15
 
 - Testing renewable subscriptions. Manual Subs have been disabled for now. Please test these. The Monthly sub only lasts for 5min. The Yearly sub will last for 1 hour.